+++ DataLayer/ViewModels/Products/ProductSearchViewModel.cs
@@ -9,7 +9,7 @@ namespace DataLayer.ViewModels.Products
         public int? Id { get; set; }
         public string Title { get; set; }
 
-        public decimal? Price { get; set; }
+        public long? Price { get; set; }
 
         public int? ProductGroupId { get; set; }
